backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
python-multipart>=0.0.17,<1
pydantic>=2.11,<3
httpx>=0.28,<1
google-genai>=1.0,<2
pypdf>=5.1,<6
pytest>=8.3,<9
